requirements.txt
fastapi==0.73.0
pandas==1.4.0
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
requests==2.27.1
SQLAlchemy==1.4.31
uvicorn==0.17.4
passlib==1.7.4